+++ src/components/Dashboard.tsx
@@ -8,6 +8,7 @@ import { PlayerGuide } from './PlayerGuide';
 import { ObjectiveBar } from './ObjectiveBar';
 import { LiveFeed } from './LiveFeed';
 import { AlertList } from './AlertList';
+import { AudioToggle } from './AudioToggle';
 
 interface Props {
   data: AllGameData;
@@ -21,11 +22,14 @@ export function Dashboard({ data }: Props) {
 
   return (
     <div className="dash">
-      {alerts.length > 0 && (
-        <div className="dash-alerts">
-          <AlertList alerts={alerts} />
-        </div>
-      )}
+      <div className="dash-topbar">
+        {alerts.length > 0 ? (
+          <div className="dash-alerts">
+            <AlertList alerts={alerts} />
+          </div>
+        ) : <div />}
+        <AudioToggle />
+      </div>
 
       <ObjectiveBar timers={timers} />